WIP - Segment refactoring - Move building operators and date to separate classes
