+++ app/bundles/LeadBundle/Segment/LeadSegmentFilterFactory.php
@@ -16,12 +16,38 @@ use Mautic\LeadBundle\Entity\LeadList;
 class LeadSegmentFilterFactory
 {
     /**
+     * @var LeadSegmentFilterDate
+     */
+    private $leadSegmentFilterDate;
+
+    /**
+     * @var LeadSegmentFilterOperator
+     */
+    private $leadSegmentFilterOperator;
+
+    public function __construct(LeadSegmentFilterDate $leadSegmentFilterDate, LeadSegmentFilterOperator $leadSegmentFilterOperator)
+    {
+        $this->leadSegmentFilterDate     = $leadSegmentFilterDate;
+        $this->leadSegmentFilterOperator = $leadSegmentFilterOperator;
+    }
+
+    /**
      * @param LeadList $leadList
      *
      * @return LeadSegmentFilters
      */
     public function getLeadListFilters(LeadList $leadList)
     {
-        return new LeadSegmentFilters($leadList);
+        $leadSegmentFilters = new LeadSegmentFilters();
+
+        $filters = $leadList->getFilters();
+        foreach ($filters as $filter) {
+            $leadSegmentFilter = new LeadSegmentFilter($filter);
+            $this->leadSegmentFilterOperator->fixOperator($leadSegmentFilter);
+            $this->leadSegmentFilterDate->fixDateOptions($leadSegmentFilter);
+            $leadSegmentFilters->addLeadSegmentFilter($leadSegmentFilter);
+        }
+
+        return $leadSegmentFilters;
     }
 }
